Fix crash with -O2/3 when collecting in main
When using -O2/3 the GC would crash if you collected in main. This was
because the gc_collect() call would get inlined, so stack_size would get
a negative value, since no new stack frame was allocated. This would
make the program segfault since we would scan way off out of the stack.

By marking it as nonline we force a stack-frame to be allocated even
when using -O3.