GC: Look for gc heap ptrs in registers, in addition to the stack
When compiling with -O2/3 I started seeing stange behavior and random
crashes. Eventually I realized this was because when compiling with
optimizations the compiler will try to avoid putting things on the stack
and keep them in registers when ever it can. This lead to those objects
not being marked, and thus discarded.

So my assumptions that all roots of the reference tree can be found on
the stack was incorrect, they can also be in registers! Now all programs
I have tried work when compiling with -O2/3.