Implement a static vtable, instead of per object
In my first implementation I had head object keep track and set up it's
own vtable. This causes the vtables to have O(n) memory complexity
instead of O(1), where n is the number of objects. This meant a Cons
object in the QuickSort file took 96 bytes, after this commit it's only
24!

In my LinkedList.p0 this lead to a ~70% speed up.