Rename constructor functions to avoid collisions
For a class A I previously called the constructor 'punkt0_new_A'. This
would conflict if a class 'new' had a method called 'A'. Unlikely maybe,
but should be avoided.

Now we do two underscores, 'punkt0_A__new', which protects against identifier collisions
since an identifier cannot start with an underscore.